requirements.txt
# Framework
Django==5.2.4

# Database
psycopg2-binary==2.9.10
dj-database-url==2.1.0

# Configuration
django-environ==0.12.0

# Testing
pytest==8.4.2
pytest-django==4.11.1
factory-boy==3.3.3
Faker==37.6.0

# Development tools
django-debug-toolbar==6.0.0

# Production Server
gunicorn==21.2.0
whitenoise==6.6.0

# Image Processing
Pillow==10.4.0